Added more delete functionality

+++ src/main/java/com/mellora/drugconsortium/controller/EmployeeController.java
@@ -1,11 +1,13 @@
 package com.mellora.drugconsortium.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -28,14 +30,40 @@ public class EmployeeController {
 		eRepo.save(employee);
 		return employee;
 	}
-	
-	@DeleteMapping(path = "/employee/remove")
-	public Employee removeEmployee(@RequestParam(name = "eFirstName") String eFirstName, @RequestParam(name = "eLastName") String eLastName) {
-		Employee employee = eRepo.findEmployeeByFirstAndLastName(eFirstName, eLastName);
-		eRepo.delete(employee);
+
+	@DeleteMapping(path = "/employee/delete")
+	public Optional<Employee> removeEmployeeByName(@RequestParam(name = "eFirstName") String eFirstName,
+			@RequestParam(name = "eLastName") String eLastName) {
+		Optional<Employee> employeeRecord;
+		if(eRepo.findEmployeeByFirstAndLastName(eFirstName, eLastName).isPresent()) {
+			employeeRecord = eRepo.findEmployeeByFirstAndLastName(eFirstName, eLastName);
+			Employee employee = employeeRecord.get();
+			eRepo.delete(employee);;
+		} else {
+			employeeRecord = null;
+		}
+		return employeeRecord;
+	}
+
+	@DeleteMapping(path = "/employee/delete/{eId}")
+	public Optional<Employee> removeEmployeeById(@PathVariable(name = "eId") long eId) {
+		Optional<Employee> employee;
+		if (eRepo.findById(eId).isPresent()) {
+			employee = eRepo.findById(eId);
+			eRepo.deleteById(eId);
+		} else {
+			employee = null;
+		}
 		return employee;
 	}
 	
+	@DeleteMapping(path = "employee/delete/all")
+	public List<Employee> removeAllEmployees(){
+		List<Employee> employeeList = eRepo.findAll();
+		eRepo.deleteInBatch(employeeList);
+		return employeeList;
+	}
+
 	@PatchMapping(path = "/employee/update")
 	public Employee updateEmployee(Employee employee) {
 		return employee;
